requirements.txt
torch
torchvision
fastapi
uvicorn
numpy
opencv-python
pillow